Fix bug

+++ HPT-OnlineCourse/src/main/java/iuh/course/hpt/repository/EnrollmentRepository.java
@@ -3,5 +3,5 @@ package iuh.course.hpt.repository;
 import iuh.course.hpt.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EnrollmentRepository extends JpaRepository<User, Integer> {
+public interface EnrollmentRepository extends JpaRepository<User, Long> {
 }
